vim: substitute handles multibyte characters
And is now in its own file
