+++ crates/vim/src/test.rs
@@ -98,27 +98,3 @@ async fn test_buffer_search(cx: &mut gpui::TestAppContext) {
         assert_eq!(bar.query_editor.read(cx).text(cx), "jumps");
     })
 }
-
-#[gpui::test]
-async fn test_substitute(cx: &mut gpui::TestAppContext) {
-    let mut cx = VimTestContext::new(cx, true).await;
-
-    // supports a single cursor
-    cx.set_state(indoc! {"ˇabc\n"}, Mode::Normal);
-    cx.simulate_keystrokes(["s", "x"]);
-    cx.assert_editor_state("xˇbc\n");
-
-    // supports a selection
-    cx.set_state(indoc! {"a«bcˇ»\n"}, Mode::Visual { line: false });
-    cx.simulate_keystrokes(["s", "x"]);
-    cx.assert_editor_state("axˇ\n");
-
-    // supports multiple cursors
-    cx.set_state(indoc! {"a«bcˇ»deˇfg\n"}, Mode::Normal);
-    cx.simulate_keystrokes(["s", "x"]);
-    cx.assert_editor_state("axˇdexˇg\n");
-
-    cx.set_state(indoc! {"ˇabc\n"}, Mode::Normal);
-    cx.simulate_keystrokes(["2", "s", "x"]);
-    cx.assert_editor_state("xˇc\n");
-}
